lang-portal/backend/requirements.txt
flask==2.3.3
flask-smorest==0.42.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pytest==7.4.2
black==23.7.0
flake8==6.1.0
